[GH-62] - better naming for analyzers
