+++ tests/NSubstitute.Analyzers.Tests.Shared/DiagnosticAnalyzers/IConflictingArgumentAssignmentsDiagnosticVerifier.cs
@@ -1,0 +1,17 @@
+using System.Threading.Tasks;
+
+namespace NSubstitute.Analyzers.Tests.Shared.DiagnosticAnalyzers
+{
+    public interface IConflictingArgumentAssignmentsDiagnosticVerifier
+    {
+        Task ReportsDiagnostic_When_AndDoesMethod_SetsSameArgument_AsPreviousSetupMethod(string method, string call, string previousCallArgAccess, string andDoesArgAccess);
+
+        Task ReportsNoDiagnostics_When_AndDoesMethod_SetsDifferentArgument_AsPreviousSetupMethod(string method, string call, string andDoesArgAccess);
+
+        Task ReportsNoDiagnostics_When_AndDoesMethod_AccessSameArguments_AsPreviousSetupMethod(string method, string call, string argAccess);
+
+        Task ReportsNoDiagnostics_When_AndDoesMethod_SetSameArguments_AsPreviousSetupMethod_SetsIndirectly(string method);
+
+        Task ReportsNoDiagnostic_When_AndDoesMethod_SetArgument_AndPreviousMethod_IsNotUsedWithCallInfo(string method, string call, string andDoesArgAccess);
+    }
+}
